+++ Sources/TableViewModel.swift
@@ -94,7 +94,16 @@ extension TableCellViewModel {
 /// that want to provide edit actions.
 public protocol TableViewCellModelEditActions {
 
+    /// The edit actions for this cell when swiping from the leading direction.
+    @available(iOS 11.0, *)
+    var leadingSwipeActionConfiguration: UISwipeActionsConfiguration? { get }
+
+    /// The edit actions for this cell when swiping from the trailing direction.
+    @available(iOS 11.0, *)
+    var trailingSwipeActionConfiguration: UISwipeActionsConfiguration? { get }
+
     /// The row edit actions for the cell.
+    /// - Warning: `UITableViewRowAction` is deprecated in iOS 13. Implement `leadingSwipeActionConfiguration` or `trailingSwipeActionConfiguration` instead.
     var editActions: [UITableViewRowAction] { get }
 }